Chatのファイル構成を変更
