+++ src/components/Chat/internal/ChatBox/index.tsx
@@ -1,0 +1,1 @@
+export { ChatBox } from './ChatBox'
