+++ src/components/Chat/internal/ChatBox/internal/ChatMessage/index.tsx
@@ -1,0 +1,1 @@
+export { ChatMessage } from './ChatMessage'
