+++ src/components/Chat/internal/ChatBox/styled.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+import { Box as MaterialUIBox } from '@material-ui/core'
+
+export const Box = styled(MaterialUIBox)`
+  height: 400px;
+  background: rgba(255, 255, 255, 0.7);
+  margin-bottom: 5px;
+`
