+++ src/components/Chat/internal/ChatMessageForm/index.tsx
@@ -1,0 +1,1 @@
+export { ChatMessageForm, ChatMessageRequest } from './ChatMessageForm'
