+++ src/components/Chat/internal/ChatMessageForm/internal/ReactButton/index.tsx
@@ -1,0 +1,1 @@
+export { ReactButton } from './ReactButton'
