+++ src/components/Chat/internal/ChatMessageForm/internal/ReactButton/styled.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import { Button } from '@material-ui/core'
+
+export const ReactButton = styled(Button)`
+  margin-right: 5px;
+`
